+++ src/platform/mod.rs
@@ -7,4 +7,5 @@ pub mod home;
 pub mod hwid;
 pub mod lock;
 pub mod mmap;
+pub mod random;
 pub mod sysinfo;